src/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.7
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
attrs==24.2.0
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
crypto==1.4.1
cryptography==43.0.3
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-login==1.10.2
fastapi-users==14.0.0
fastapi-users-db-sqlalchemy==6.0.1
filelock==3.16.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
identify==2.6.2
idna==3.10
isort==5.13.2
Jinja2==3.1.4
makefun==1.15.6
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
Naked==0.1.32
nodeenv==1.9.1
passlib==1.7.4
platformdirs==4.3.6
pre-commit==3.8.0
propcache==0.2.1
psycopg2-binary==2.9.10
pwdlib==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyTelegramBotAPI==4.24.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
ruff==0.7.1
shellescape==3.8.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
starlette==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.27.1
yarl==1.18.3
